Add minified dependency on fastutil
Fastutil (https://github.com/vigna/fastutil) is an excellent
library for memory-efficient and fast collections. Gradle has
many algorithms that could benefit from these more efficient
data structures.

Since the library is huge (16MB), this change introduces a dependency
transformation that can minify external JAR files to only the classes
we need. This could be useful for shrinking the Gradle distribution
in general.

This commit only adds the classes necessary to make IntOpenHashSet
work, which will be used in a subsequent commit to make our incremental
Java compiler more memory efficient.
