+++ subprojects/base-services/base-services.gradle
@@ -13,6 +13,7 @@ sourceCompatibility = javaVersion.java9Compatible ? 1.6 : 1.5
 dependencies {
     api libraries.guava
     api libraries.jsr305
+    api libraries.fastutil
 
     implementation libraries.slf4j_api
     implementation libraries.commons_lang
