+++ subprojects/messaging/src/main/java/org/gradle/internal/serialize/IntSetSerializer.java
@@ -1,0 +1,4 @@
+package org.gradle.internal.serialize;
+
+public class IntSetSerializer {
+}
